% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resautomodel.R
\name{hiddenBlock}
\alias{hiddenBlock}
\title{hiddenBlock}
\usage{
hiddenBlock(inlayer, nodes, acts, idepth, orginlayer = NULL,
  reg = NULL, dropout = 0, batchnorm = TRUE)
}
\arguments{
\item{inlayer}{input layer, keras layer}

\item{nodes}{list of integers, list of the number of nodes for all the hidden layers}

\item{acts}{list of strings, list of activations for hidden layers}

\item{idepth}{integer, index of the hidden layer}

\item{orginlayer}{keras layer, original layer to be added to decoding layer (default: NULL)}

\item{reg}{string, regularization (default: NULL)}

\item{dropout}{double, dropout rate for the target hidden layer (default 0)}

\item{batchnorm}{bool, flag to conduct batch normalization (default: TRUE)}
}
\value{
keras layer, block of a hidden layer (with addtion of actionvation or/and batch normalization)
}
\seealso{
 \code{\link{AutoEncoderModel}} for major function using this function.
}
\description{
This function is to construct the hidden block .
}
