% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{keras_r2}
\alias{keras_r2}
\title{keras_r2}
\usage{
keras_r2(y_true, y_pred)
}
\arguments{
\item{y_true}{tensor of ground truth}

\item{y_pred}{tensor of predicted values}
}
\value{
keras tensor, double for rsquared
}
\description{
This function is to calculate rsquared value for regression models.
}
