% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{r2_squ}
\alias{r2_squ}
\title{r2_squ}
\usage{
r2_squ(obs, res)
}
\arguments{
\item{obs}{vector, observed values}

\item{res}{residual (observed values-predicted values)}
}
\value{
double for rsquared
}
\description{
This function is to calculate rsquared value for regression models.
}
\examples{
n=300
x=runif(n,1,100)
y=3*x^2+sqrt(abs(x))+rnorm(n)
lmodel=lm(y~x)
ypre=predict(lmodel,data=x)
r2=r2_squ(y,y-ypre)
message(paste("r2 is :",r2))
}
