% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rmse}
\alias{rmse}
\title{rmse}
\usage{
rmse(obs, pre)
}
\arguments{
\item{obs}{vector, observed values}

\item{pre}{vector, predicted values}
}
\value{
double for rmse
}
\description{
This function is to calculate rmse value for regression models.
}
\examples{
n=200
x=runif(n,1,10)
y=log(3*x)+x^2+rnorm(n)
lmodel=lm(y~x)
ypre=predict(lmodel,data=x)
rmse=rmse(y,ypre)
message(paste("rmse is :",rmse))
}
