% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rSquared}
\alias{rSquared}
\title{rSquared}
\usage{
rSquared(obs, res)
}
\arguments{
\item{obs}{vector, observed values}

\item{res}{residual (observed values-predicted values)}
}
\value{
double for rSquared
}
\description{
This function is to calculate rsquared value for regression models.
}
