% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaleautonet.R
\name{ResautoDownscale}
\alias{ResautoDownscale}
\title{ResautoDownscale}
\usage{
ResautoDownscale(r2, fpredict0, c.grid, ss = 0.2, nepoch = 30,
  cores = 1, thresh = 0.01, ntime = 5)
}
\arguments{
\item{r2}{: Stack for the covariates for downscaling}

\item{fpredict0:}{Starting predictions obtained by GWR or other optimal learners such as XGBoost, residual network}

\item{c.grid:}{Coarsely resolved grid}

\item{ss:}{sampling propotion for independent test}

\item{nepoch:numder}{of epoch for residual network training (default: 30)}

\item{corse:}{Number of CPU cores used for computing (default: 1)}

\item{cores:}{double, dropout rate for the target hidden layer (default 0)}

\item{thresh:}{stopping creation shreshold (default: 0.01)}

\item{ntime:}{maximum number of iterations (default: 5)}
}
\value{
List(performance metrics such as R2, RMSE, and downscaled images)
}
\description{
This function is the iterative implementation of downscaling with autoencoder based residual network.
}
